feat: 커스텀 구조변경
